+++ NEWS.md
@@ -1,5 +1,6 @@
 # ocetag 0.0.3
 
+* [ctdTagApp()] can now plot sigma-spiciness diagrams.
 * Database internal version changed from 1 to 2, and the `level`
   element of the `tags` table was renamed as `index`.
 * Add [readDatabase()], which may be handy for users who do not know